Fixed the issue of workers being added and removed for each comparison, which resulted in the functions being compiled for each comparison significantly slowing down the parallel code. This fix adjusts when the cores are added and cleans up all workers after every analysis is run. This only impacts the additional cores and not the master core.

+++ inst/jl/Euclidean_Distance_Operations.jl
@@ -3,9 +3,9 @@
 #############################################
 #Three-dimensional
 @everywhere function MD3D(v1, m2)
-	dsum = Inf
+	dsum::Float64 = Inf
 	for j in 1:size(m2,1)
-		dtemp = sqrt((v1[1] - m2[j,1]) * (v1[1] - m2[j,1]) + (v1[2] - m2[j,2]) * (v1[2] - m2[j,2]) + (v1[3] - m2[j,3]) * (v1[3] - m2[j,3]))
+		dtemp::Float64 = sqrt((v1[1] - m2[j,1]) * (v1[1] - m2[j,1]) + (v1[2] - m2[j,2]) * (v1[2] - m2[j,2]) + (v1[3] - m2[j,3]) * (v1[3] - m2[j,3]))
 		if dtemp < dsum
 			dsum = dtemp
 		end
@@ -15,7 +15,7 @@ end
 
 #Two-dimensional
 @everywhere function MD2D(v1, m2)
-	dsum = Inf
+	dsum::Float64 = Inf
 	for j in 1:size(m2,1)
 		dtemp = sqrt((v1[1] - m2[j,1]) * (v1[1] - m2[j,1]) + (v1[2] - m2[j,2]) * (v1[2] - m2[j,2]))
 		if dtemp < dsum
@@ -30,7 +30,7 @@ end
 #############################################
 #Three-dimensional
 @everywhere function AD3D(v1, m2)
-	dsum = zeros(1,size(m2,1))
+	dsum::Float64 = zeros(1,size(m2,1))
 	for j in 1:size(m2,1)
 		dsum[1,j] = sqrt((v1[1] - m2[j,1]) * (v1[1] - m2[j,1]) + (v1[2] - m2[j,2]) * (v1[2] - m2[j,2]) + (v1[3] - m2[j,3]) * (v1[3] - m2[j,3]))
 	end
@@ -39,7 +39,7 @@ end
 
 #Two-dimensional
 @everywhere function AD2D(v1, m2)
-	dsum = zeros(1,size(m2,1))
+	dsum::Float64 = zeros(1,size(m2,1))
 	for j in 1:size(m2,1)
 		dsum[1,j] = sqrt((v1[1] - m2[j,1]) * (v1[1] - m2[j,1]) + (v1[2] - m2[j,2]) * (v1[2] - m2[j,2]))
 	end
@@ -88,7 +88,7 @@ end
 #############################################
 #Three-dimensional
 function MED3D(m1, m2)
-	n1 = size(m1,1)
+	n1::Int = size(m1,1)
 	Dist = SharedArray{Float64}(n1)
 	@sync @distributed for k in 1:n1
 		Dist[k] = MD3D(m1[k,:], m2)
@@ -98,7 +98,7 @@ end
 
 #Two-dimensional
 function MED2D(m1, m2)
-	n1 = size(m1,1)
+	n1::Int = size(m1,1)
 	Dist = SharedArray{Float64}(n1)
 	@sync @distributed for k in 1:n1
 		Dist[k] = MD2D(m1[k,:], m2)
@@ -111,8 +111,8 @@ end
 #############################################
 #Three-dimensional
 function EDM3D(m1, m2)
-	n1 = size(m1,1)
-	n2 = size(m2,1)
+	n1::Int = size(m1,1)
+	n2::Int = size(m2,1)
 	Dist = SharedArray{Float64}(n1,n2)
 	@sync @distributed for k in 1:n1
 		Dist[k,:] = AD3D(m1[k,:], m2)
@@ -122,8 +122,8 @@ end
 
 #Three-dimensional
 function EDM2D(m1, m2)
-	n1 = size(m1,1)
-	n2 = size(m2,1)
+	n1::Int = size(m1,1)
+	n2::Int = size(m2,1)
 	Dist = SharedArray{Float64}(n1,n2)
 	@sync @distributed for k in 1:n1
 		Dist[k,:] = AD2D(m1[k,:], m2)
@@ -136,7 +136,7 @@ end
 #############################################
 #Three-dimensional
 function MEDI3D(m1, m2)
-	n1 = size(m1,1)
+	n1::Int = size(m1,1)
 	Dist = SharedArray{Float64}(n1,2)
 	@sync @distributed for k in 1:n1
 		Dist[k,:] = MDI3D(m1[k,:], m2)
@@ -146,7 +146,7 @@ end
 
 #Two-dimensional
 function MEDI2D(m1, m2)
-	n1 = size(m1,1)
+	n1::Int = size(m1,1)
 	Dist = SharedArray{Float64}(n1,2)
 	@sync @distributed for k in 1:n1
 		Dist[k,:] = MDI2D(m1[k,:], m2)
